requirements.txt
opencv-python>=4.5.0
mediapipe>=0.8.9
numpy>=1.19.0
pyserial>=3.5